Fixed things, broke other things, resolved msot of the comments from the PR however I'm passing a tuple where it should be an int.
