+++ app/controllers/admin/userManagement.py
@@ -4,10 +4,9 @@ from app.controllers.admin import admin_bp
 from app.models.user import User
 from app.models.program import Program
 from app.logic.userManagement import addCeltsAdmin,addCeltsStudentStaff,removeCeltsAdmin,removeCeltsStudentStaff
-from app.logic.userManagement import changeCurrentTerm
 from app.logic.userManagement import changeProgramInfo
 from app.logic.utils import selectSurroundingTerms
-from app.logic.userManagement import addNextTerm
+from app.logic.term import addNextTerm, changeCurrentTerm
 from app.models.term import Term
 
 @admin_bp.route('/admin/manageUsers', methods = ['POST'])
