+++ app/controllers/main/routes.py
@@ -36,7 +36,7 @@ from app.logic.manageSLFaculty import getCourseDict
 from app.logic.courseManagement import unapprovedCourses, approvedCourses
 from app.logic.utils import selectSurroundingTerms
 from app.logic.certification import getCertRequirementsWithCompletion
-from app.logic.serviceLearningCoursesData import pushDataToDatabase
+from app.logic.serviceLearningCoursesData import pushDataToDatabase, sessionCleaner
 from app.logic.createLogs import createRsvpLog, createAdminLog
 
 @main_bp.route('/logout', methods=['GET'])
@@ -391,10 +391,8 @@ def getAllCourseInstructors(term=None):
         dataHolder = []
         termHolder = []
 
-    if 'errorFlag' in session:
-        errorFlag = session['errorFlag']
-    else:
-        errorFlag = ''
+    errorFlag = session.get('errorFlag')
+
 
     if g.current_user.isCeltsAdmin:
         setRedirectTarget("/manageServiceLearning")
@@ -407,10 +405,7 @@ def getAllCourseInstructors(term=None):
 
         if request.method =='POST' and "submitParticipant" in request.form:
             pushDataToDatabase(session['dataPreview'])
-            session.pop('dataPreview')
-            session.pop('listofBnumber_students')
-            session.pop('errorFlag')
-            session.pop('termDict')
+            sessionCleaner()
             dataHolder =[]
             flash('File saved successfully!', 'success')
             return redirect(url_for('main.getAllCourseInstructors'))
